remove testfiles package
