+++ internal/playwright/config.go
@@ -12,7 +12,6 @@ import (
 	"github.com/saucelabs/saucectl/internal/fpath"
 	"github.com/saucelabs/saucectl/internal/msg"
 	"github.com/saucelabs/saucectl/internal/region"
-	"github.com/saucelabs/saucectl/internal/testfiles"
 )
 
 // Config descriptors.
@@ -224,7 +223,7 @@ func shardInSuites(rootDir string, suites []Suite, ccy int) ([]Suite, error) {
 			return []Suite{}, err
 		}
 
-		testFiles := testfiles.ExcludeFiles(files, excludedFiles)
+		testFiles := fpath.ExcludeFiles(files, excludedFiles)
 
 		if s.Shard == "spec" {
 			for _, f := range testFiles {
