+++ internal/testcafe/config.go
@@ -13,7 +13,6 @@ import (
 	"github.com/saucelabs/saucectl/internal/fpath"
 	"github.com/saucelabs/saucectl/internal/msg"
 	"github.com/saucelabs/saucectl/internal/region"
-	"github.com/saucelabs/saucectl/internal/testfiles"
 )
 
 // Config descriptors.
@@ -263,7 +262,7 @@ func Validate(p *Project) error {
 				return err
 			}
 
-			p.Suites[i].Src = testfiles.ExcludeFiles(files, excludedFiles)
+			p.Suites[i].Src = fpath.ExcludeFiles(files, excludedFiles)
 		}
 	}
 
@@ -315,7 +314,7 @@ func shardSuites(rootDir string, suites []Suite, ccy int) ([]Suite, error) {
 			return []Suite{}, err
 		}
 
-		testFiles := testfiles.ExcludeFiles(files, excludedFiles)
+		testFiles := fpath.ExcludeFiles(files, excludedFiles)
 
 		if s.Shard == "spec" {
 			for _, f := range testFiles {
